Update Ledger deps
The latest version of `@ledgerhq/hw-transport-u2f` (and `hw-transport`) appear to have shipped a strange mix of transpiled and untranspiled code, and they break the local build. This is extremely hacky, but since the build is account viewer build is so modularized and fragile, I found it easiest to vendor the deps and transform the code manually. This is disgusting and I hate it